refactor: 중복 기능 클래스 제거
